requirements.txt
itemadapter==0.4.0
pymongo==4.3.3
Scrapy==2.7.1
Twisted==22.1.0
psycopg2-binary==2.9.7
redis==4.5.4
celery==5.3.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0
fake-useragent==1.4.0
scrapy-user-agents==0.1.1
scrapy-rotating-proxies==0.6.2
scrapy-splash==0.8.0
scrapy-redis==0.7.3
scrapy-crawlera==1.1.0
scrapy-zyte-smartproxy==1.2.0
pandas==2.1.1
numpy==1.24.3
pytz==2023.3
schedule==1.2.0
python-dateutil==2.8.2
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
flask-login==0.6.3
flask-wtf==1.1.1
wtforms==3.0.1
werkzeug==2.3.7
bcrypt==4.0.1
email-validator==2.0.0
flask-admin==1.6.1
flask-migrate==4.0.5
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.21
